name_resolver write state for empty arrays
Summary:
This diff implements the name-resolver side of things for .push providers for empty arrays. It doesn't yet update the new_env so there isnt an external effect.

We create a new write state for empty arrays, and a new kind of assignment in the env_api for empty arrays as well. These states contain the locations of any push providers that the provider_analysis has determined will affect the type of the array.

We also record in the env_state the locations of push providers themselves, which will be used to populate a corresponding slot in the loc_env.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37078571

fbshipit-source-id: 67c7b8220b77161f4da121a3172491bae75c179d