Find providers for .push, indexed array writes, and array writes to empty arrays
Summary:
This diff expands on the provider analysis to handle empty arrays as a special case, which will reduce the need to add annotations to them.

At present, a variable like `var x = []` is regarded as fully "provided for" by `[]`, even though this expression (in the absence of a contextual type) says little about the actual type of the variable. In order to have meaningful behavior on it, we would need to annotate it when we switch to LTI--and some cases, where the variable is written to again later, we have already required annotations for due to constrained writes.

This diff changes the behavior of such variables by allowing them to widen in two ways. First of all, if a variable is initialized to [], we now allow the first subsequent write of a non-empty array literal to widen the type. So `var x = []; x = [1]` will mean that `x` has type `Array<number>`. This is analogous to how `null`-initialized variables are allowed to widen by their first write.

The second change is that now the first call of .push on an empty array, or the first indexed write to it, also now are counted as providers, and we'll (in subsequent diffs) use these expressions to initialize the array's type as well. Since these are, conceptually, writes to the array's element tvar rather than the array itself, we need to treat them differently, keeping them separate in the provider_api from the actual direct writes to the variable.

Reviewed By: SamChou19815

Differential Revision: D37059538

fbshipit-source-id: f86338bf6ea414ab9906eb220157c3ca06517f56