Typecheck empty arrays based on providers
Summary:
This diff finishes the implementation of array element providers from .push and indexed writes.

The main contribution of this diff is the handling of the `EmptyArrayWrite` kind during env initialization. Rather than just creating an empty tvar as normal, we actually create the overall array type here, since we have everything we need to synthesize it here already. We'll look up the tvars for array providers, which are now stored in a loc_env map similarly to the environment itself, and create a type representing them. This type will then be a lower bound on the element type of an ArrT, which we store in the environment. This provider type will also be a post-inference upper bound on the type of the ArrT's element type tvar, in order to detect illegal writes to the array.

Unfortunately, this post-inference check can produce bad error messages, because post-inference resolution of tvars erases the use_ops from the tvars lower bounds. For example, if we had a flow like
```
NumT --AssignVar--> OpenT --UnknownUse--> StrT
```
We would normally get a good error message like "Cannot assign ... because number is incompatible with string". But if we resolve the OpenT and flow it post-inference to StrT via the UnknownUse use_op, we'll just get "number is incompatible with string." This situation is very relevant in this diff, because we don't have a use_op for the constraint on the array's tvar. We can reconstruct an ok erorr message in some cases by tweaking how the post-inference check resolution happens: when the post-inference LB is a tvar, we explore it directly and resolve all of its lower bounds and flow them to the upper bound directly, rather than resolving the LB tvar. However, this doesn't help for cases where the error is deeper, for example if we push an `Array<string>` to an `Array<Array<number>>`. For this I think we just have to live with the bad error message until LTI is complete, at which point the distinction will hopefully go away and we won't lose use_ops.

The other aspect of this diff worth calling out is that during checking, we now always check whether an expression is an array provider, and store in in the loc_env if so.

Reviewed By: SamChou19815

Differential Revision: D37081649

fbshipit-source-id: c84d02b274e4947aba77f053212058f37ea93fb8