Internal module names in new_env
Summary: This diff builds support for declared modules in the new env (which seem to be the only way that `InternalModuleNames` exist in the environment). There's not much to this--a declared module name points at itself in the new_env, and is written to via the normal Env_sig entry points used by statement.ml.

Reviewed By: SamChou19815

Differential Revision: D36985101

fbshipit-source-id: bbc1e67c79c0fa52f1f8cc12f7cc930087c60f4b