[new-env] Prevent missing provider crash when the provider is a non-assigning write
Summary:
Right now under new-env, we crashes under the following code with reasons in comments:

```
let x = 0;
let x: number = 0; // provider of x
x = 0; // subtype_against_provider will try to read the slot of the provider, which doesn't exist. CRASH!
```

Currently, the provider analysis assumes that your code has no illegal redeclarations. (While this is a strong assumption and definitely doesn't hold in general, it's pretty hard to accidentally write code like above, that's why we have never caught the problem). For declarations in the same scope, like the two lets in the example above, will be considered as provider candidates. If we end up choosing the first `let` as the provider, then we won't crash, but it chooses the second `let` in this case because it has an annotation, and it causes a crash.

In general, when there are illegal declarations in the same scope, and we choose the one different from the one that will be visited first in `name_resolver`, so the provider loc is a `NonAssigningWrite`, then we can potentially run into these crashes.

This diff fixed the above mentioned problem by creating a slot on loc_env for non-assigning providers.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D37154086

fbshipit-source-id: 450512df2916b5862a798e84f04845ea85d35cca