[resolved-env] Do not generate a global mutable node_cache
Summary:
Changelog: [internal]

In the current implementation, `Node_cache.empty` is an immutable module-level value with mutable fields. It is used as the initial empty value for context. This means that even if we reset the context to the initial empty value, we are still reading the same globally mutable cache!

This is causing all sorts of crashes in IDE tests, because they usually make more than one request to a long-live server. When we type check the same file the second time for IDE purposes, we will still read the incorrectly cached value that might have a different tvar graph.

This diff ensures that we generate a new empty value for each `mk_empty` call, so that other code can use the usual mechanism to reset to an empty state.

Reviewed By: samwgoldman

Differential Revision: D37440710

fbshipit-source-id: f0bf13b121eea6e8c19bd440ce7b0d6b4b8b70af