Make object_widening test rely less on weird flow behavior
Summary:
A lot of funky stuff happens here. First of all, widening is order-specific, and secondly there is definitely some funkiness happening in here with the behavior that adopts properties from an upper bound in object literal lower bounds.

Removing the cast and adding a spread_exact fixes a lot of the weird behaviors, and more closely aligns the behavior here with what we expect in the resolved env.

The next diff re-records the remaining missing duplicate errors for the fully resolved env.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37461603

fbshipit-source-id: 93cb87504c728d6b083c848755a9eb7af1107a68