[resolved-env] Correctly report missing local annotations in resolved-env
Summary:
Changelog: [internal]

The only place we can possibly report missing local annotations is the `Contextual` def in fully resolved environment. If we don't have a hint for that contextual def, we will error.

Right now, this is not handled correctly in env_resolution. Even before contextual typing is done, we can't completely rely on statement.ml to report these errors, because we might have

```
const f = (x) => ...;
```

which will produce a `Value` def for `f`. Once we resolved the RHS, it will be cached forever, so statement.ml doesn't even have a chance to visit the unannotated `x` parameter. Therefore, we must perform the detection in env_resolution as well.

The detection itself is similar to the existing one in statement.ml. I only need to slightly tweak the contextual def to pass the reason of the missing annotation pattern, so that the error messages can match existing ones.

Now that we have ensured that we can detect all cases of `missing-local-annot` error, we can pass dummy hints to all other callers of `Statement.expression` in `env_resolution`, so we don't produce spurious missing local annotations error. This is safe since these annotations are not used for producing any types, only their existence matters for now. In the future, they can be replaced with `Hint_None` once the hint logic has been completely removed from statement.ml

Reviewed By: mvitousek

Differential Revision: D37369239

fbshipit-source-id: d2e2e1b1815ab716e77c1b749fbd7983479ebc80