[resolved-env] Re-record test suites with minor global read error message location differences
Summary:
Changelog: [internal]

We only error on using unresolvable global once, at the location when it's first visited. Resolved-env will change the error we visit the code, so sometimes it will not error on the first global read. This difference is not very important, so let's just re-record the tests.

Reviewed By: samwgoldman

Differential Revision: D37500864

fbshipit-source-id: e10811eeb01843abfd7782cbc466505575c0f1c5