don't store warnings in saved state
Summary:
Warnings should not be stored in saved state. They are found during check, and in lazy mode should only be reported on focused files. When a file gets focused, we check it, so will find the warnings then. When they are in saved state, they get reported even on unfocused files!

This also fixes a crash when loading a saved state that contains warnings with cross-module references. The other files haven't been parsed yet, so their ALoc tables in shared mem are missing, leading to an exception.

Changelog: [fix] Fix a bug where warnings would be incorrectly reported when using saved state, and could cause a crash.

Reviewed By: jbrown215

Differential Revision: D37521354

fbshipit-source-id: 475a5c7c3fbf54b55237a4c1aec60779b8870bac