[resolved-env] Allow unify_declared_type to unify even when the entry is marked as resolved
Summary:
Changelog: [internal]

In resolved-env, we will mark the slots of the defs to be resolved, which prevents further unification happening in the main pass of statement.ml. This makes sense since the types are already resolved, and there is no need to add in further constrains.

However, there are some functions that currently attempt to unify with the loc_env slot not for the purpose of adding constrains, but for the purpose of getting these constrains. For example, `unify_declared_type` serves to adds the annotation constrains to the loc env slot when there is an annotation, but it will get the constrain from the loc env when there is no annotation. Very often, the type we passed to `unify_declared_type` will be the type we stick into the AST, which will affect type-at-pos results. If we fail to unify, the type-at-pos result will return empty.

Reviewed By: jbrown215

Differential Revision: D37479493

fbshipit-source-id: 74f3b7e7757951dfed4852dbda8b321704ffd426