Consistently wrap destructured variable in AnnotT
Summary:
Changelog: [internal]

Right now destructuring is not done in a consistent manner, which poses a challenge to fully resolved environement where all the destructuring is done centrally in env_resolution in quite a uniform matter.

This diff fixes an easy problem regarding protecting destructured variables deriving from an annotation with AnnotT. This is currently done inconsistently. It was done in both `destructuring.ml` (which is the place it should be done), and in additional in statement.ml during variable declaration, with the same kind of flow but slightly different reason.

I found the flow used in statement.ml to be slightly better, because with that flow, the reason will be pointing to the annotation when it exists, so I changed `destructuring.ml` to use this flow and removed the now unnecessary flow in statement.ml

Reviewed By: samwgoldman

Differential Revision: D37566815

fbshipit-source-id: 9af58ff629da286375a3c6a5a8b0ec2aff3ef5f4