[EZ] Kill statement-level reordering flag
Summary:
Changelog: [internal]

We will not use statement level reordering for LTI. Instead, we will use a more fine-grained order produced by the env_builder to do reordering to expression levels. Therefore, this never-advertised flag would no longer make sense, and this diff deletes it and all of the associated statement reordering code.

A cleanup of toplevel.ml logic will be done in the next diff

Reviewed By: samwgoldman

Differential Revision: D37536209

fbshipit-source-id: 0ef117825b9d1d6b7bc00041ff91c04687cabde0