Drop illegal unbinding from enums tests
Summary:
Enum methods must be called with the correct `this`, so it's unsafe to unbind
these functions and then call them.

Currently, function types do not constrain `this`, so casting these unbound
methods to this-less function types passes. In the next diff, I will make
function types strict w.r.t. `this`, causing these lines to fail.

Splitting this change out to make the next diff slightly easier to review.

Reviewed By: SamChou19815

Differential Revision: D37582673

fbshipit-source-id: 82b6d759d503164a14d9843d1d670fad2740fe76