Refactor: key the environment maps, and related maps used in constructing the environment, by kind of binding as well as location
Summary:
This diff introduces a major refactoring of the name resolver, loc env, and name def analysis.

### The problem
Previously, the loc env and name resolver contained a number of distinct ALoc maps representing different kinds of data: ordinary variables, and the different varieties of `this` and `super`. In addition, the ordinary variables map actually contained more than "just" ordinary variables: it also contained entries for arbitrary expressions that are used to compute types for variables, e.g. the right hand side of an `instanceof` test. So far we've been careful to ensure that we would never have a collision between two different "conceptual" entries in these maps (e.g. a variable definition, and an auxiliary expression used in computing variable types), but it's always been a risk. When we would have such a collision, we introduced a new map, like the `this` maps; this works, but increases the complexity of the data structures.

### This diff
This diff addresses this risk by changing the structure of the loc env and name resolver maps from being a collection of locmaps to being a single map called an EnvMap, which is keyed by the pair of a location and a Env_api.name_def_type. This means that a single map contains all the entries, and there isn't a risk of overlap as long as the same location can be distinguished by name_def_type. In addition, we introduce name_def_types for auxiliary expressions and for array providers, in order to ensure that they never collide with each other or with ordinary variables.

In addition--simply because this diff is a convenient time to make this change--the loc env now stores a bare Type.t rather than an annotated_or_inferred.

Reviewed By: SamChou19815

Differential Revision: D37631410

fbshipit-source-id: 340e6ccf84881e5e960b6625e5383d74092d705b