fix parser exception on negative number types
Summary:
number types can be negative, and can have whitespace between the `-` and the number.

we parse octal (`0x...`) and binary (`0b...`) numbers by calling `Int64.of_string` on them, but that throws an exception if they contain whitespace.

to fix, we need to strip out the whitespace. this would be easy if sedlex supported named capture groups but it doesn't :/

Changelog: [parser] Fix parser exception on negative binary and octal number types that contain whitespace after the minus sign

Reviewed By: SamChou19815

Differential Revision: D37756160

fbshipit-source-id: c44ef25524810a4477896715e13b07fa69463a26