LTI codemod now annotates class properties and method returns when enforce_class_annotations enabled
Summary:
This diff extends the LTI codemod to handle the missing annotation errors introduced in D37741137. This change is relatively straightforward, since the errors are of the same kind as other local inference errors--we simply need to check if there is an LTI error pointing at the new cases that we care about. Specifically, we now need to see if there are missing annotation errors pointing at initialized class properties and at function return types (for methods). If there are, we annotate them in the same way we do elsewhere. These errors are only added if the `enforce_class_annotations` flowconfig option D37739269 is enabled, so we don't need to check for that option here--if the option isn't enabled, the errors that trigger annotation insertion won't occur in the first place.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D37739269

fbshipit-source-id: 1b0da5b34f1682978cfb329d564e1874acd2e7f0