fix extra line when autoimporting
Summary:
if an autoimport is inserted at the bottom of a section and there's another section below it, we incorrectly include an extra blank line.

we scan the existing imports looking for the best place to insert the new import. when we find it, we previously inserted the new one above the next import. this is a problem when the next import is in a new section and already has a blank line above it.

now we insert the new import below the previous import in this case.

Changelog: [fix] Fix inconsistent blank lines around auto-imports

Reviewed By: samwgoldman

Differential Revision: D37861656

fbshipit-source-id: 96811deb7dd279f0318db32cd46d5bf371e7baca