[resolved-env] Flow function type and class type to general to provide type-at-pos result
Summary:
Changelog: [internal]

The `general` type of functions and classes in this case is the type that will be stored in the AST, to later power generics escape check and type-at-pos queries. Right now for inferred functions and classes in which we will cache the AST and never visit it again, we need to ensure that we get the type of `general` correct. Currently, these `general` types will receive no lower bound, so they will all be printed as `empty`.

This diff flows the function and class type into the general type to provide it with a lower bound. This allows us to produce correct type-at-pos result as before.`

Reviewed By: mvitousek

Differential Revision: D37500410

fbshipit-source-id: 93a4cbde684c04042b52bdd14517c98225748e57