[resolved-env] Do not add ChainExpression def if it is the callee of optional call
Summary:
Changelog: [internal]

For new-env, we generate a write for refinement that's never written to in optional calls like `foo.bar?.(...)`. We are treating `foo.bar` as a write, so that we can later read a refined version of it. However, this write is never used in any way in statement.ml. In statement.ml, this is implemented without out any recording of write of `foo.bar` or refinement read using `Refinements.get`. Instead, it's implemented using tvars. Currently under resolved-env, we will attempt to resolve this def, and we will type check `foo.bar` in isolation. This causes a bunch of problems, including incorrect use_op and additional method unbinding errors.

In this diff, I choose to solve the problem for the purpose of resolved-env rollout, by not registering them as defs for resolved-env. We can also avoid generating such writes in name_resolver, but I believe these writes should be there, and in the long term, we should make statement.ml actually write to and read from the slot when we disentangle method type checking.

Reviewed By: mvitousek

Differential Revision: D37439601

fbshipit-source-id: af3d8ba9c5d29c39c70a8d698ffce52fca854b83