[resolved-env] Defaults in destructuring
Summary:
Changelog: [internal]

Right now defaults in destructuring are not computed correctly, but it just happens to work in most cases. The missing piece is that we need to add the following flows `type of default ~> type of destructured name` and `type of parameter default ~> type of parameter`.

To achieve this, we need to do some bookkeeping similar to what we did in destructuring.ml. Notably, we need to create a data structure similar to the one in `default.ml` that tracks how defaults concatenate and destruct as well. I did that in name_def that generates a syntactic version of it, and it will be converted to `Default.t` in env_resolution to give us the default type. This structure will be attached to all selectors, and `expression option` that represents the default expression will be attached to annotation and contextual root.

There is also an interesting behavior we need to replicate:
If the function parameter has annotations, then the type of default expression will be the parameter annotation, even if the default expression is not a subtype of the annotation. This ensures that we don't unnecessarily error in two different places, like in the following case:

```
function f({p}: {p: string} = {p: boolean}) {}
```

We achieve that by calling `Statement.Func_stmt_config.eval_default` which has the behavior, and cache the expression, so that later when we are creating `Default.t`, the same annotation type will be guaranteed to created again.

Reviewed By: mvitousek

Differential Revision:
D37573528 (https://github.com/facebook/flow/commit/0c31af36395ebcaaf9e0be113fed545cec480a9d)

------------------------------------------------------------------------
(from 4429dd140a18d89f7ff15ee8a644446186295922)

fbshipit-source-id: d02cb4f06ffd4ed1bd1f9aceae2ad587d270aff1