make MethodTs' action optional
Summary:
Changelog: [internal]

In contextual typing, we want to use the callee type to contextually type the arguments. When a call expression is a method call, we need to use `MethodT` to get the callee type. We only want the type but without "invoking" the method, so we need to make the action optional.

Reviewed By: samwgoldman

Differential Revision: D37792019

fbshipit-source-id: fad6df07b7fc0d5e6fcd4e641405011615a1f2e4