Functorize type annotation over constraint generation
Summary:
This work functorizes `type_annotation.ml` over the places where it uses constraint generation by calling into Flow_js.ml, and provides an alternative implementation of constraint generation using annotation inference. It also exposes and changes some types from annotation inference in order to align with the signature of Flow_js's version.

The version using annotation inference will crash in a lot of places so it's not ready for use, and depending on short term plans it may not be fixed soon. However we'll likely eventually need something like this, so let's put it out there now to make sure that the signatures stay in sync.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37863196

fbshipit-source-id: 9b3fa24d44f057c0b31cfbe74ed08a46492250df