Initial cut at sealing tvars in env resolution
Summary:
If the new `enforced` env mode is enabled, we can now seal tvars! This works by just exploring the types of whatever was added to the environment when typechecking a component, and transforming all tvars to `FullyResolved` ones.

There are some missing pieces here: some name_defs (classes, eg) add multiple types to the environment, so we will need to do more work to seal them, and as a result classes don't work under this mode. In addition, unsurprisingly, contextually typed functions don't work--the argument is resolved to empty.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D37888880

fbshipit-source-id: 80e2c2d02ef2431fca252deaf03f6995d6c6b976