Additional checking that components have been handled correctly
Summary:
In order to help guide implementation of tvar resolution, this diff adds additional sanity checking to the new environment. It validates 1. the only environment entries that we're allowed to read from are either part of the current component, or of components that had been scheduled earlier and are now complete, and 2. that only entries from the current component may be unresolved--entries from completed components must be FullyResolved.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37908940

fbshipit-source-id: aa4292d3e9de6b2fa0efc432ca50d244c92c6292