Improve annotate useState codemod to support null/undefined/empty-array arguments
Summary:
A call to `useState` with just a `null`, `undefined`, or `[]` needs to be annotated (along with `{}`). Add support for this to the codemod. Also adds a `--generalize-maybe` flag which may be useful for the `null`/`undefined` arguments.

Currently in WWW:
- `useState(null)`: 3912
- `useState(undefined)`: 80
- `useState([])`: 2059

`new Map()` and `new Set()` will be more complicated, will do in a subsequent diff.

Changelog: [feature] Improve annotate useState codemod to support null/undefined/empty-array arguments

Reviewed By: SamChou19815

Differential Revision: D37949323

fbshipit-source-id: e4cce9f3efbd0069b09da84bea0a3f07ca301fad