[hoister] Completely ignore declare module statements in hoister
Summary:
Changelog: [internal]

For

```
declare module foo {
  declare var bar: string;
}
```

We now hoist `bar` to the toplevel, which disagrees with the intended semantics of declare module scoping.

In this diff, I decide to completely ignore declare module statements. Later in the stack, I will special case dealing with declare module with its own hoister that visits the body.

Reviewed By: mvitousek

Differential Revision: D37831307

fbshipit-source-id: f1363c753dff8135d65127ba2306d0d34012f6dc