[new-env] `exclude_syms` support
Summary:
Changelog: [internal]

This is the final prep diff to support checking libdefs with new-env.

Different libdefs might define the same global names. We will type check libdefs in certain order, and the first libdef that defines a global name will win. From there, the global name will be added to `exclude_syms` set, and all other modules' definition of the name will be ignored, and their read of the name will lead the type of the first definition.

In name_resolver, we will model this behavior in terms of read and write. When a hoisted binding's name is in the `exclude_syms` set, it will not be added to the environment. Similarly, the name also cannot be written, so we need to perform conditional checks in places like `pattern_identifier` and `binding_type_identifier`. With these two checks, reads on those excluded names will naturally point to a global read.

Reviewed By: panagosg7

Differential Revision: D37869327

fbshipit-source-id: c1c88565761072a07e7017bb52010b57e9e2e9a8