handle React Class Components in post-inference implicit instantiations
Summary:
When implicitly instantiating a type parameter on a class component we trust the variance annotation on the class.

In practice, the only valid variance here would be Invariant, since the type argument to React.Component is invariant.

An improvement here may be to find the props object by traversing the inheritance chain and then being more precise about how the variance of the type parameters actually appear in the Props object.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D26315603

fbshipit-source-id: 63e15723a3799c49a4978e55242056ff426a3e3c