[resolved-env] Fully resolve the env entry tvar after unification
Summary:
Changelog: [internal]

In new_env.ml, we will error in enforced mode when a readable entry is not under resolution and not fully resolved:

https://github.com/facebook/flow/blob/2fedca00a07e434d4f03eaf3d1ae10b8870f55ce/src/typing/new_env.ml#L183-L201

This check makes sense, but we might fail to make entries a `FullyResolved` tvar even when it's actually fully resolved. When we computed the potentially not fully resolved type `t` for an env entry in env resolution, the following two things happen:

1. `t` is unified with an OpenT.
2. `t` becomes fully resolved by a call to `TvarResolver.resolve`

In step 1, if `t` is not an OpenT (e.g. when `t` is an annotation), then for the OpenT env entry, it will be `OpenT(reason, id)` where id points to `Resolved(t)` instead of `FullyResolved(lazy t)`.

To fix the problem, we can fully resolve the env entry instead of `t`. This will always ensure that the env entry points to a `FullyResolved` constraint.

Reviewed By: jbrown215

Differential Revision: D38042554

fbshipit-source-id: 1988026a87cd34a7b0cdb9922e19630d40192252