[new-env] Use new-env to type check libdefs
Summary:
Changelog: [internal]

In this diff, we will finally use the new-env to type check libdefs. Currently, there are two old-env specific ways in which libdefs are checked.

1. For global names, old-env reads the module-level scope entries.
2. For names in declare module, old-env will read the scope of declare module in a declare module without any exports.

Of course, the above methods cannot work in the new-env. Instead, the new-env will use the results of name_resolver to obtain the types:

1. We find the global names by hoister, and then we simulate a read on them. We will get a list of `(name, read)` tuple.
2. For each declare module, we also simulate a read for each toplevel names. In the end, we will also get a list of `(name, read)` tuple.

Both lists will be properly filtered according to `exclude_syms`.

In new-env, the reads will be translated to types, and added to global or export map.

Reviewed By: panagosg7

Differential Revision: D37736292

fbshipit-source-id: 583899189e3cb06a2095e8245255bd01cdbbb3a9