Collate Import Index
Summary:
Update Export_search with correct counts for each export by using the information from imports_to_add and imports_to_remove. Now, the environment's export_search's count for an export equal to (the number of times that the export is imported + 1) and will go up and down as you add and remove imports. Currently, when you remove an export and re-add it, the export will start with a count of 1and changes based on any imports you add or remove from then on, which is not ideal(should be addressed in the future).

Changelog: [internal] Collating import index so export_search has correct counts

Reviewed By: SamChou19815

Differential Revision: D38011798

fbshipit-source-id: 54ecafbf5b835df11e1514493f97fc0c5b7cf9d9