Make transaction commit/rollback non-cooperative
Summary:
Before this change, each commit/rollback operation returned a Lwt.t. In
practice, these functions would never yield. However, on transaction
commit/rollback, we would yield between individual operations due to Lwt.iter_s.

This diff moves the commit/rollback operations out of Lwt and also ensures that
we never yield between operations.

Finally, I removed the "singleton" API. This is a well-intentioned API, but the
use case has evolved and it's not really relevant anymore. It is still dangerous
to create the same mutator multiple times in a single transaction, but IMO it is
not worthwhile to protect against this at runtime.

Reviewed By: mroch

Differential Revision: D38212469

fbshipit-source-id: 717907cb6d7ff12e2ce2f38699e5d68b46bc1135