Handle class this/super during resolution of classes, and seal afterwards
Summary:
The various forms of `this` and `super` need to be constructed during resolution of a class signature during env resolution and sealed afterwards. This diff changes the point at which they're bound (in the new env only) to `mk_class_sig`. It also in the env resolver introduces the ability to have more than one entry point to tvar resolution--the this and supers may not be directly reachable from a class type, but they must nonetheless be resolved in the same component as the class.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37914475

fbshipit-source-id: dfcafb45ce6543ef912de10186cd306c8aa5ced4