Construct class types during signature resolution
Summary:
After constructing a class sig, we have all the information we need about it to construct its types--we don't need to wait until the class's body is checked (under the new restrictions on class behavior). We should do so so that those types can be sealed at the end of checking the component!

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37915008

fbshipit-source-id: 4ba72704367bd3625c389b1b2122d18e59714b60