Split function and global this and position global this at program loc
Summary:
Currently global this is located at ALoc.none. In a later diff, I'm adding some internal errors that point to environment entries, and this causes a crash if an error points at the entry for global this. In order to fix this, this diff positions global this at the program loc. However, this can overlap with a function this loc if the entire module is a single function, so this diff also separates the global this kind from the function this kind so prevent them from colliding in the env.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37909136

fbshipit-source-id: 784f647e09d0e14c907a896084cea2c1fd478c95