Error on missing writes
Summary:
This diff adds enforcement to make sure that environment entries and the name resolver are synced up. When we read from the loc env and find nothing at a given location, either  it's a loc that happens to have been "populated" by a `NonAssigningWrite`, which is fine, or it's an improper loc that should not have been queried. This adds an error in the latter case so we can go and fix the query. The error is gated under the Enforced env mode.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37913950

fbshipit-source-id: 03995493eb744226384a3bac5fece8a893199417