Remove special casing of `this` in ordering and handle `this` annotations
Summary:
The name_def_ordering previously special-cased `this` to detect recursive classes (e.g. most of them). This is no longer necessary, plus `this` is now handled in the env builder. This diff removes most special casing of `this` and instead builds direct support for `this` annotations (as well as support for `this` more generally, which should fall directly out from name resolver behavior).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37913957

fbshipit-source-id: d0db864254035e6d789d7b0dcbceb9dd2fcd69b2