Clear entire check contents cache on commit
Summary:
The check contents cache stores merged type information for dependencies.
Merging type information is somewhat expensive and we call check contents quite
frequently for IDE commands.

Check contents reads committed data, which can change when a transaction
commits. Due to laziness, the cache contains not only types, but also lazy
thunks that close over shared heap addresses.

If we do not invalidate the cache correctly, we might get out-of-date type
information, but even worse we might try to access stale heap addresses. We
clear this cache during compact, so values will not have moved, but they might
have been swept and marked as free.

Before this diff, we tried to be clever and only removed "sig_new_or_changed"
files. In principle, this should be enough. There isn't a known issue with this
invalidation strategy. However, getting invalidation wrong is scary enough that
we should be more conservative.

This diff takes a more heavy-handed approach and just clears the entire cache
after a recheck. This means IDE commands that immediately succeed a recheck may
get a lower cache hit rate, but the cache will re-populate quickly and
subsequent commands will be fast.

Reviewed By: mroch

Differential Revision: D38248896

fbshipit-source-id: 5054ad2eb02674a5e32b53ddcd9caf0d42d35fd2