Make a solve_targs function for use in the enforced environment
Summary:
It seems pretty clear that we'll need a function that can solve for type arguments and return a substitution map. Most other plans we have around the results of implicit instantiation can be implemented given this output.

Since we need to return a type for constant tparams, this diff also adds the pinned constant tparam to the callback and calls the missing bounds callback when no bounds are available. We can likely relax that behavior by adding another callback for constant tparams with no bounds, but that is out of the scope of this diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38082194

fbshipit-source-id: 39db35f06fa31fbb4b8c2b3a5e810c9b577a4b71